PdfReader/requirements.txt
camelot_py
opencv_python
